conversation refactor
